Update yubi-mode.sh, README
Added a fix to yubi-mode.sh that can check usbguard rules and add the key if it's being blocked.

This was a common cause of the card being 'stuck' in an unreadable state where the light would continuously blink.

+++ gnupg/yubi-mode.sh
@@ -9,13 +9,47 @@ RESET="\033[00m"
 # Install this script to your preferred location, with the correct permissions (root:root, 755) and add it to your PATH.
 
 # Troubleshooting:
+#
+# If you're passing the yubikey through to a Linux VM from a Linux Host with usbguard, the yubikey must be allowed in the host's
+# usbguard rules for the VM to read it, even if the VM can 'see' it as a device.
+#
 # Sometimes the yubikey can become 'stuck' in a state preventing gpg operations with the card, and the light on the card will continuously blink.
 # This can happen when changing keys with different gpg identities.
-# The quickest fix is logging out, log back in, then re-running this script in gpg mode.
+# The quickest fix is to log out, log back in, then run this script again in gpg mode.
 
 if [ "${EUID}" -eq 0 ]; then
 	echo "Run this as a normal user. Sudo is used to elevate only specific commands."
 	exit
+fi
+
+if (systemctl is-active --quiet usbguard); then
+#	if (sudo tail /var/log/usbguard/usbguard-audit.log | grep -Pq "^.+result='SUCCESS'.+target.new='block'.+name \"YubiKey .+$"); then
+	if (sudo usbguard list-devices | grep -Pq "^\d+: block id \d{4}:\d{4} serial \"\" name \"YubiKey .+$"); then
+
+		echo -e "${YELLOW}[i]usbguard service is blocking the YubiKey...${RESET}"
+
+#		echo -e "${BLUE}[>]Remove the YubiKey...${RESET}"
+#		while (lsusb | grep -q 'Yubico.com'); do
+#			sleep 1
+#		done
+
+#		echo -e "${BLUE}[>]Reconnect the YubiKey...${RESET}"
+#		until (sudo usbguard list-devices | grep -Pq "^\d+: block id \d{4}:\d{4} serial \"\" name \"YubiKey .+$"); do 
+#			sleep 1
+#		done
+
+# OPTION 1, less reliable
+#		DEVICE_ID="$(sudo usbguard list-devices | grep -P "^\d+: block id \d{4}:\d{4} serial \"\" name \"YubiKey .+$" | grep -oP "^\d+")"
+#		sudo usbguard allow-device "$DEVICE_ID"
+
+# OPTION 2, most reliable
+		ALLOW_RULE="$(sudo usbguard list-devices | grep -P "^\d+: block id \d{4}:\d{4} serial \"\" name \"YubiKey .+$" | sed 's/^[[:digit:]]\{1,3\}: block/allow/')"
+		echo "$ALLOW_RULE" | sudo tee -a /etc/usbguard/rules.conf > /dev/null
+		sudo systemctl restart usbguard
+		echo -e "${BLUE}[✓]Device allowed.${RESET}"
+	fi
+else
+	echo -e "${YELLOW}[i]Error modifying usbguard rules.${RESET}"
 fi
 
 if [ "${1}" == 'gpg' ]; then
@@ -23,9 +57,11 @@ if [ "${1}" == 'gpg' ]; then
         echo -e "${BLUE}[>]GPG mode${RESET}"
 	if (pgrep yubioath > /dev/null); then
 		echo -e "${YELLOW}[i]Stopped runnning yubioath-desktop...${RESET}"
-		pkill yubioath
+		pkill -f 'yubioath-desktop'
 	fi
-        sudo systemctl stop snap.yubioath-desktop.pcscd.service
+	if (systemctl is-active --quiet snap.yubioath-desktop.pcscd.service); then
+	        sudo systemctl stop snap.yubioath-desktop.pcscd.service
+	fi
         sudo systemctl restart pcscd
 
 	# https://github.com/drduh/YubiKey-Guide#switching-between-two-or-more-yubikeys
@@ -38,6 +74,11 @@ if [ "${1}" == 'gpg' ]; then
         echo -e "${BLUE}[✓]Done.${RESET}"
 
 elif [ "${1}" == 'otp' ]; then
+	if ! (command -v yubioath-desktop); then
+		echo -e "${YELLOW}[i]yubioath-desktop application not found. Quitting.${RESET}"
+		exit 1
+	fi
+
 	# Connects yubikey to the yubioath-desktop snap application to view OTP codes
         echo -e "${BLUE}[>]OTP mode${RESET}"
 
@@ -45,8 +86,10 @@ elif [ "${1}" == 'otp' ]; then
 	pkill gpg-agent ; pkill ssh-agent ; pkill pinentry
 	#eval $(gpg-agent --daemon --enable-ssh-support)
 
-        sudo systemctl stop pcscd
-        sudo systemctl restart snap.yubioath-desktop.pcscd.service
+	if (systemctl is-active --quiet pcscd); then
+		sudo systemctl stop pcscd
+	fi
+	sudo systemctl restart snap.yubioath-desktop.pcscd.service
 	echo -e "${YELLOW}[i]Starting yubioath-desktop as background process...${RESET}"
 	yubioath-desktop &
